+++ CHANGELOG.md
@@ -8,8 +8,9 @@
 - GP2-1426 - Cost and pricing - Change to the title of the tool-tip
 - GP2-1424 - Cost and pricing page - local taxes section
 - GP2-1423 - Cost and pricing page - content change
-- GP2-1411 - Allow type of article in ArticlePage to be blank, not just null
+- GP2-1411 - Refactor ArticlePage to use a StreamField with new PullQuoteBlock
 - GP2-1333 - Add filtering behaviour to MarketsTopicLandingPage (ported from BAU)
+- GP2-1412 - Update RichtextField features allowed on ArticlePage
 - GP2-1402 - Update content for Funding and Credit section
 - GP2-1407 - Rename AdviceTopicLandingPage to just TopicLandingPage
 - GP2-1338 - Port v1 ArticlePage template + model